+++ car-service-server/car-service-web/src/main/java/com/woniu/car/service/web/controller/UserCarServiceController.java
@@ -3,6 +3,7 @@ package com.woniu.car.service.web.controller;
 
 import com.woniu.car.commons.core.code.ConstCode;
 import com.woniu.car.commons.core.dto.ResultEntity;
+import com.woniu.car.commons.core.exception.CarException;
 import com.woniu.car.items.model.entity.UserCarService;
 import com.woniu.car.items.model.param.carservice.DeleteCarServiceParam;
 import com.woniu.car.items.model.param.userservice.AddUserServiceParam;
@@ -13,6 +14,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.BeanUtils;
+import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
@@ -44,12 +46,15 @@ public class UserCarServiceController {
     @ApiOperation(value = "新增用户服务关联信息",notes = "flag为true时新增成功,false时生成失败")
     @ApiImplicitParam(name = "addUserServiceParam",value = "接收新增用户服务关联信息参数",required = true,dataType = "AddUserServiceParam")
     public ResultEntity addUserCarService(@RequestBody AddUserServiceParam addUserServiceParam){
+        if (ObjectUtils.isEmpty(addUserServiceParam)){
+            return ResultEntity.buildFailEntity().setMessage("参数为空").setFlag(false).setCode(ConstCode.LAST_STAGE);
+        }
         UserCarService userCarService = new UserCarService();
         BeanUtils.copyProperties(addUserServiceParam,userCarService);
         System.out.println("UserCarService"+":"+userCarService);
         boolean b = userCarServiceService.addUserCarService(userCarService);
         if (b) return ResultEntity.buildSuccessEntity().setMessage("新增成功").setCode(ConstCode.ACCESS_SUCCESS);
-        return ResultEntity.buildFailEntity().setMessage("新增失败").setCode(ConstCode.LAST_STAGE).setFlag(false);
+        throw new CarException("新增失败",500);
     }
     /**
      * @Author HuangZhengXing
@@ -62,10 +67,12 @@ public class UserCarServiceController {
     @ApiOperation(value = "根据用户id查询关联表信息",notes = "flag为true时新增成功,false时生成失败")
     @ApiImplicitParam(name = "listUserServiceByUserParam",value = "接收要查询关联表信息的用户id",required = true,dataType = "ListUserServiceByUserParam")
     public ResultEntity<List<UserCarService>> listUserServiceByUser(@RequestBody ListUserServiceByUserParam listUserServiceByUserParam){
+        if (ObjectUtils.isEmpty(listUserServiceByUserParam.getUserId())) return ResultEntity.buildListEntity(UserCarService.class).setMessage("输入为空，或输入有误").setFlag(false).setData(null);
         UserCarService userCarService = new UserCarService();
         BeanUtils.copyProperties(listUserServiceByUserParam,userCarService);
         System.out.println("UserCarService"+":"+userCarService);
         List<UserCarService> userCarServices = userCarServiceService.listUserCarServiceByUser(userCarService);
+        if (ObjectUtils.isEmpty(userCarService)) throw new CarException("结果为空",500);
         return ResultEntity.buildListSuccessEntity(UserCarService.class).setMessage("查询成功").setCode(ConstCode.ACCESS_SUCCESS).setData(userCarServices);
     }
     /**
@@ -79,6 +86,7 @@ public class UserCarServiceController {
     @ApiOperation(value = "查询所有用户服务关联信息",notes = "不需要携带参数")
     public ResultEntity<List<UserCarService>> listUserCarServiceAll(){
         List<UserCarService> userCarServiceList = userCarServiceService.listUserCarServiceAll();
+        if (ObjectUtils.isEmpty(userCarServiceList)) return ResultEntity.buildListSuccessEntity(UserCarService.class).setMessage("查询成功,结果为空").setCode(ConstCode.ACCESS_SUCCESS).setData(userCarServiceList);
         return ResultEntity.buildListSuccessEntity(UserCarService.class).setMessage("查询成功").setCode(ConstCode.ACCESS_SUCCESS).setData(userCarServiceList);
     }
 
@@ -93,12 +101,17 @@ public class UserCarServiceController {
     @ApiOperation(value = "根据用户服务关联id修改服务状态",notes = "flag为true时新增成功,false时生成失败")
     @ApiImplicitParam(name = "updateUserCarServiceStatusParam",value = "接收要修改关联表信息id和要修改的状态",required = true,dataType = "UpdateUserCarServiceStatusParam")
     public ResultEntity updateUserServiceStatus(@RequestBody UpdateUserCarServiceStatusParam updateUserCarServiceStatusParam){
+        if (ObjectUtils.isEmpty(updateUserCarServiceStatusParam.getUserServiceId())){
+            return ResultEntity.buildListEntity(UserCarService.class).setMessage("输入为空，或输入有误").setFlag(false).setData(null);
+        }else if(ObjectUtils.isEmpty(updateUserCarServiceStatusParam.getUserServiceStatus())){
+            return ResultEntity.buildListEntity(UserCarService.class).setMessage("输入为空，或输入有误").setFlag(false).setData(null);
+        }
         UserCarService userCarService = new UserCarService();
         BeanUtils.copyProperties(updateUserCarServiceStatusParam,userCarService);
         System.out.println("UserCarService"+":"+userCarService);
         boolean b = userCarServiceService.updateCarServiceStatus(userCarService);
         if (b) return ResultEntity.buildSuccessEntity().setMessage("修改成功").setCode(ConstCode.ACCESS_SUCCESS);
-        return ResultEntity.buildFailEntity().setMessage("修改失败").setCode(ConstCode.LAST_STAGE).setFlag(false);
+        throw new CarException("修改失败",500);
     }
 
     /**
@@ -112,12 +125,13 @@ public class UserCarServiceController {
     @ApiOperation(value = "根据用户服务关联id删除信息",notes = "flag为true时新增成功,false时生成失败")
     @ApiImplicitParam(name = "deleteCarServiceParam",value = "接收要删除的关联表id",required = true,dataType = "DeleteCarServiceParam")
     public ResultEntity deleteUserService(@RequestBody DeleteCarServiceParam deleteCarServiceParam){
+        if (ObjectUtils.isEmpty(deleteCarServiceParam.getCarServiceId())) return ResultEntity.buildListEntity(UserCarService.class).setMessage("输入为空，或输入有误").setFlag(false).setData(null);
         UserCarService userCarService = new UserCarService();
         BeanUtils.copyProperties(deleteCarServiceParam,userCarService);
         System.out.println("UserCarService"+":"+userCarService);
         boolean b = userCarServiceService.deleteCarService(userCarService);
         if (b) return ResultEntity.buildSuccessEntity().setMessage("删除成功").setCode(ConstCode.ACCESS_SUCCESS);
-        return ResultEntity.buildFailEntity().setMessage("删除失败").setCode(ConstCode.LAST_STAGE).setFlag(false);
+        throw new CarException("删除失败",500);
     }
 }
